my_pro/requirements.txt
crewai>=0.28.0
torch>=2.0.0
transformers>=4.35.0
PyYAML>=6.0
accelerate>=0.20.0
sentencepiece>=0.1.99
protobuf>=3.20.0
